These vars are supposed to be specific size.  Assume our systems are up to date with common mid-90's practice and have specific size types available, and use them.  Previous had bad results on e.g. common 64-bit systems wiereh 'long' is not in fact 32 bits.

+++ util.c
@@ -3943,8 +3943,9 @@ void ConstrainByBorders (TwmWindow *twmwin,
 
 #ifdef JPEG
 
-unsigned short int *buffer_16bpp;
-long *buffer_32bpp;
+/* May need stdint.h (on C99) or inttypes.h (on not-quite) systems? */
+uint16_t *buffer_16bpp;
+uint32_t *buffer_32bpp;
 
 static void convert_for_16 (int w, int x, int y, int r, int g, int b) {
   buffer_16bpp [y * w + x] = ((r >> 3) << 11) + ((g >> 2) << 5) + (b >> 3);
